% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_compound.R
\name{create_compound_map}
\alias{create_compound_map}
\title{Create a compound mapping}
\usage{
create_compound_map()
}
\value{
Returns a list of functions to operate with the mapping, see \code{\link{create_maptype_map}}.
}
\description{
Create a compound mapping by merging the effect of individual mappings.
}
\details{
The following fields are required in the parameter list to initialize the mapping:
\tabular{ll}{
\code{mapname} \tab Name of the mapping \cr
\code{maptype} \tab Must be \code{"compound_map"} \cr
\code{maps} \tab a list with the parameter lists of the individual maps
}

This mapping will also take care of the correct ordering of the individual
mappings.
Bundling together individual mapping specifications to a compound map is
the last step before doing Bayesian inference.
}
\examples{
params1 <- list(
  mapname = "mymap1",
  maptype = "linearinterpol_map",
  src_idx = 1:3,
  tar_idx = 7:8,
  src_x = 1:3,
  tar_x = 2:3
)
params2 <- list(
  mapname = "mymap2",
  maptype = "linearinterpol_map",
  src_idx = 4:6,
  tar_idx = 7:8,
  src_x = 1:3,
  tar_x = 2:3
)
compmap_params <- list(
  mapname = "mycompmap",
  maptype = "compound_map",
  maps = list(params1, params2)
)
mycompmap <- create_compound_map()
mycompmap$setup(compmap_params)
x <- c(1,2,3,5,5,5,0,0)
mycompmap$propagate(x)
mycompmap$jacobian(x)


}
\seealso{
Other mappings: 
\code{\link{create_convolution_with_xtrafo_map}()},
\code{\link{create_derivative2nd_map}()},
\code{\link{create_derivative_map}()},
\code{\link{create_linearinterpol_map}()},
\code{\link{create_linearinterpol_with_xtrafo_map}()},
\code{\link{create_linearmap_map}()},
\code{\link{create_maptype_map}()},
\code{\link{create_map}()},
\code{\link{create_nonlinear_map}()},
\code{\link{create_normerr_map}()},
\code{\link{create_product_map}()}
}
\concept{mappings}
