% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_derivative.R
\name{create_derivative_map}
\alias{create_derivative_map}
\title{Create a mapping from function values to the first derivative}
\usage{
create_derivative_map()
}
\value{
Returns a list of functions to operate with the mapping, see \code{\link{create_maptype_map}}.
}
\description{
Create a mapping from a discretized version of a function with values
given on an x-mesh to the first derivatives for a subset of the x-values
of the source mesh.
}
\details{
The following fields are required in the parameter list to initialize the mapping:
\tabular{ll}{
\code{mapname} \tab Name of the mapping \cr
\code{maptype} \tab Must be \code{"derivative_map"} \cr
\code{src_idx} \tab Vector of source indices \cr
\code{tar_idx} \tab Vector of target indices. \cr
\code{src_x} \tab Vector of x-values of the source mesh \cr
\code{tar_x} \tab Vector of x-values of the target mesh.
                  Must be a subset of the x-values of the source mesh
                  and must not include the largest x-value
                  of the source mesh.
}

\loadmathjax
Let \mjseqn{x_i} denote the x-values of the mesh and \mjseqn{y_i} the
associated y-vales of the function. A finite-difference approximation of
the first derivative is given by
\mjsdeqn{
  \Delta_i = \frac{\vec{y}_{i+1} - \vec{y}_i}{\vec{x}_{i+1}-\vec{x}_i}
}
}
\examples{
params <- list(
  mapname = "myderivmap",
  maptype = "derivative_map",
  src_idx = 1:4,
  tar_idx = 5:6,
  src_x = 1:4,
  tar_x = 2:3
)
mymap <- create_derivative_map()
mymap$setup(params)
x <- c(1,2,4,7,0,0)
mymap$propagate(x)
mymap$jacobian(x)

}
\seealso{
Other mappings: 
\code{\link{create_compound_map}()},
\code{\link{create_convolution_with_xtrafo_map}()},
\code{\link{create_derivative2nd_map}()},
\code{\link{create_linearinterpol_map}()},
\code{\link{create_linearinterpol_with_xtrafo_map}()},
\code{\link{create_linearmap_map}()},
\code{\link{create_maptype_map}()},
\code{\link{create_map}()},
\code{\link{create_nonlinear_map}()},
\code{\link{create_normerr_map}()},
\code{\link{create_product_map}()}
}
\concept{mappings}
