% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_linearmap.R
\name{create_linearmap_map}
\alias{create_linearmap_map}
\title{Create a generic linear mapping}
\usage{
create_linearmap_map()
}
\value{
Returns a list of functions to operate with the mapping, see \code{\link{create_maptype_map}}.
}
\description{
Create a linear mapping between the variables at the source indices
to the variables at the target indices, which is of the form
\mjsdeqn{
  \vec{y} = \vec{y}_{\textrm{ref}} + S (\vec{p} - \vec{p}_{\textrm{ref}})
}
}
\details{
The following fields are required in the parameter list to initialize the mapping:
\tabular{ll}{
\code{mapname} \tab Name of the mapping \cr
\code{maptype} \tab Must be \code{"linearmap_map"} \cr
\code{src_idx} \tab Vector of source indices \cr
\code{tar_idx} \tab Vector of target indices. \cr
\code{yref} \tab The vector \mjseqn{\vec{y}_{\textrm{ref}}} given in the formula. \cr
\code{pref} \tab The vector \mjseqn{\vec{p}_{\textrm{ref}}} given in the formula. \cr
\code{S} \tab The matrix \eqn{S} given in the formula.
}
\loadmathjax
}
\examples{
params <- list(
  mapname = "mylinearmap",
  maptype = "linearmap_map",
  src_idx = 1:2,
  tar_idx = 3:5,
  yref = rep(10, 3),
  pref = c(2, 4),
  S = matrix(c(1, 2, 3, 4, 5, 6), nrow=3, ncol=2)
)
mymap <- create_linearmap_map()
mymap$setup(params)
x <- c(1,-2,0,0,0)
mymap$propagate(x)
mymap$jacobian(x)

}
\seealso{
Other mappings: 
\code{\link{create_compound_map}()},
\code{\link{create_convolution_with_xtrafo_map}()},
\code{\link{create_derivative2nd_map}()},
\code{\link{create_derivative_map}()},
\code{\link{create_linearinterpol_map}()},
\code{\link{create_linearinterpol_with_xtrafo_map}()},
\code{\link{create_maptype_map}()},
\code{\link{create_map}()},
\code{\link{create_nonlinear_map}()},
\code{\link{create_normerr_map}()},
\code{\link{create_product_map}()}
}
\concept{mappings}
