% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utilities.R
\name{create_map}
\alias{create_map}
\title{Create a mapping}
\usage{
create_map(params)
}
\arguments{
\item{params}{Parameter list with the mapping definition, see \strong{Details}.}
}
\value{
Returns a list of functions to operate with the mapping, see \code{\link{create_maptype_map}}.
}
\description{
This function creates a mapping from a parameter list.
}
\details{
It calls the mapping creation function responsible for mappings
of \code{maptype} stated in the parameter list. It also
invokes the \code{setup} function to initialize the mapping,
provided by all mapping objects, see the \strong{Value} section
of \code{\link{create_maptype_map}}.
For all maps the fields \code{mapname} and \code{maptype} must be present
in the parameter list.
The other parameters depend on the specifics
of the mapping, see e.g., \code{\link{create_linearinterpol_map}}.
}
\note{
This function should be preferred over the specialized mapping creation
functions, such as \code{\link{create_linearinerpol_mapping}}.
}
\examples{
params <- list(
  mapname = "mylinearintmap",
  maptype = "linearinterpol_map",
  src_idx = 1:3,
  tar_idx = 4:6,
  src_x = c(1,5,10),
  tar_x = c(4,5,6)
)
mymap <- create_map(params)
x <- c(1,2,3,0,0,0)
mymap$propagate(x)
mymap$jacobian(x)

}
\seealso{
Other mappings: 
\code{\link{create_compound_map}()},
\code{\link{create_convolution_with_xtrafo_map}()},
\code{\link{create_derivative2nd_map}()},
\code{\link{create_derivative_map}()},
\code{\link{create_linearinterpol_map}()},
\code{\link{create_linearinterpol_with_xtrafo_map}()},
\code{\link{create_linearmap_map}()},
\code{\link{create_maptype_map}()},
\code{\link{create_nonlinear_map}()},
\code{\link{create_normerr_map}()},
\code{\link{create_product_map}()}
}
\concept{mappings}
