% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nonlinear.R
\name{create_nonlinear_map}
\alias{create_nonlinear_map}
\title{Create a mapping to effect a non-linear transformation}
\usage{
create_nonlinear_map()
}
\value{
Returns a list of functions to operate with the mapping, see \code{\link{create_maptype_map}}.
}
\description{
Creates a mapping that transforms the values at the source indices by
a predefined non-linear transformation and outpts them at the target
indices. Supported transformations at present are \emph{exp}, \emph{relu}
and \emph{limiter}.
}
\details{
The following fields are required in the parameter list to initialize the mapping:
\tabular{ll}{
\code{mapname} \tab Name of the mapping \cr
\code{maptype} \tab Must be \code{"nonlinear_map"} \cr
\code{src_idx} \tab Vector of source indices \cr
\code{tar_idx} \tab Vector of target indices.
                    Must be of same length as \code{src_idx} \cr
\code{funname} \tab A string specifying the non-linear transformation.
Can be \code{"exp"}, \code{"relu"} or \code{"limiter"}. \cr
\code{minvalue} \tab If \code{funname = "limiter"}, the lower cut-off limit. \cr
\code{maxvalue} \tab If \code{funname = "limiter"}, the upper cut-off limit.
}
}
\examples{
params <- list(
  mapname = "mynonlinearmap",
  maptype = "nonlinear_map",
  src_idx = 1:3,
  tar_idx = 4:6,
  funname = "relu"
)
mymap <- create_nonlinear_map()
mymap$setup(params)
x <- c(1,-2,3,0,0,0)
mymap$propagate(x)
mymap$jacobian(x)

}
\seealso{
Other mappings: 
\code{\link{create_compound_map}()},
\code{\link{create_convolution_with_xtrafo_map}()},
\code{\link{create_derivative2nd_map}()},
\code{\link{create_derivative_map}()},
\code{\link{create_linearinterpol_map}()},
\code{\link{create_linearinterpol_with_xtrafo_map}()},
\code{\link{create_linearmap_map}()},
\code{\link{create_maptype_map}()},
\code{\link{create_map}()},
\code{\link{create_normerr_map}()},
\code{\link{create_product_map}()}
}
\concept{mappings}
