% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_optimhelp.R
\name{evaluate_logpostpdf}
\alias{evaluate_logpostpdf}
\title{Evaluate posterior density function}
\usage{
evaluate_logpostpdf(map, zprior, U, obs, zref)
}
\arguments{
\item{map}{Mapping object. Usually a compound map, see \code{\link{create_compound_map}}}

\item{zprior}{Vector of prior estimates of the independent variables (i.e., associated with nodes without parent nodes)}

\item{U}{Prior covariance matrix of the independent variables}

\item{obs}{Vector with observed values of dependent nodes. Must be of same
size as \code{zprior}. An \code{NA} value in this vector means that the
corresponding variable was not observed.}

\item{zref}{Vector with values of independent variables used as reference vector
in Taylor expansion. The posterior pdf is also evaluated exactly
at this point.}
}
\value{
Return a list with the following elements:
\tabular{ll}{
  \code{val} \tab Value of the posterior probability density function evaluated at \code{zref} \cr
  \code{jac} \tab Gradient of the posteroir probability density function evaluated at \code{zref}
}
}
\description{
Evaluate the posterior density function for a given vector of
independent variables.
}
