% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_optimhelp.R
\name{generate_logpostpdf_funs}
\alias{generate_logpostpdf_funs}
\title{Generate helper functions for optimization routines}
\usage{
generate_logpostpdf_funs(map, zprior, U, obs, zref)
}
\arguments{
\item{map}{Mapping object. Usually a compound map, see \code{\link{create_compound_map}}}

\item{zprior}{Vector of prior estimates of the independent variables (i.e., associated with nodes without parent nodes)}

\item{U}{Prior covariance matrix of the independent variables}

\item{obs}{Vector with observed values of dependent nodes. Must be of same
size as \code{zprior}. An \code{NA} value in this vector means that the
corresponding variable was not observed.}

\item{zref}{Vector with values of independent variables used as reference vector
in Taylor expansion. The posterior pdf is also evaluated exactly
at this point.}
}
\value{
Return a list with two functions:
\tabular{ll}{
  \code{fun(x)} \tab Function that evaluates the posterior pdf at \code{x} \cr
  \code{jac(x)} \tab Function that evaluates the gradient of the posterior pdf at \code{x}
}
}
\description{
Generate a list with two function to evaluate the posterior probability
density function and its gradient. These functions can be used in
optimization to find the posterior maximum.
}
