% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utilities.R
\name{get_network_structure}
\alias{get_network_structure}
\title{Get network structre}
\usage{
get_network_structure(maplist, nodes, obs, nonlinear_ind = TRUE)
}
\arguments{
\item{maplist}{A list of mapping objects}

\item{nodes}{A vector with the node name associated to the variables.
Several variables can belong to the same node.}

\item{obs}{A vector of observed values. Unobserved values are represented by
\code{NA}.}

\item{nonlinear_ind}{Should non-linear interactions between nodes be
indicated by dashed lines.}
}
\value{
Return an igraph graph object
}
\description{
Create a Bayesian network graph
}
\examples{
\dontrun{
  library(igraph)
  params1 <- list(
    mapname = "mylinearmap1",
    maptype = "linearinterpol_map",
    src_idx = 1:3,
    tar_idx = 4:6,
    src_x = c(1,5,10),
    tar_x = c(4,5,6)
  )
  params2 <- list(
    mapname = "mylinearmap2",
    maptype = "linearinterpol_map",
    src_idx = 4:6,
    tar_idx = 7:9,
    src_x = c(4,5,7),
    tar_x = c(4.5, 5, 5.5)
  )
  compparams <- list(
    mapname = "mycompmap",
    maptype = "compound_map",
    maps = list(params1, params2)
  )

  mymap1 <- create_map(params1)
  mymap2 <- create_map(params2)

  mymap <- create_map(compparams)
  maplist <- mymap$getMaps()
  nodes <- c(rep("mynode1", 3),
             rep("mynode2", 3),
             rep("obsnode", 3))
  obs <- rep(NA, 9)

  grph <- get_network_structure(maplist, nodes, obs)
  plot(grph)
}

}
