% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucdataBaynet-package.R
\docType{package}
\name{nucdataBaynet}
\alias{nucdataBaynet}
\title{nucdataBaynet - Nuclear data evaluation with Bayesian networks}
\description{
Bayesian networks with multivariate priors on variables and
linear and non-linear relationships between variables.
}
